add mobile version of app
